Cleaned up code, moved cpu_init to cpu.cpp
